+++ src/Adjustments/Contracts/Adjustable.php
@@ -18,13 +18,10 @@ use Illuminate\Database\Eloquent\Relations\MorphMany;
 
 interface Adjustable
 {
-    /** @todo rename this in v4. In case of a cart item or and order item, this name is just stupid */
-    public function itemsTotal(): float;
+    public function preAdjustmentTotal(): float;
 
-    /**
-     * @todo add this to the interface in v4
-     * public function invalidateAdjustments(): void
-     */
+    public function invalidateAdjustments(): void;
+
     public function adjustments(): AdjustmentCollection;
 
     public function adjustmentsRelation(): MorphMany;